Use double instead of float for intermediate random number computation

+++ pythran/pythonic/include/numpy/random/weibull.hpp
@@ -12,12 +12,12 @@ namespace numpy
   namespace random
   {
     template <class pS>
-    types::ndarray<double, pS> weibull(float a, pS const &shape);
+    types::ndarray<double, pS> weibull(double a, pS const &shape);
 
-    auto weibull(float a, long size)
+    auto weibull(double a, long size)
         -> decltype(weibull(a, types::array<long, 1>{{size}}));
 
-    double weibull(float a, types::none_type size = {});
+    double weibull(double a, types::none_type size = {});
 
     DEFINE_FUNCTOR(pythonic::numpy::random, weibull);
   }
